arch: Refactor architecture library
This commit splits the "architecture" concept into two, more specific
concepts: the vendor, and the architecture. This allows us to refactor
architecture support into more common parts, and contain
architecture-specific logic more cleanly.

Some architectural functions and definitions have also been renamed to
better reflect their origin and to maintain naming consistency.

As architectures are generally not intended to be built independently of
any firmware, the CI script no longer tries to build the architecture
library.
