+++ arch/arm/armv7-m/src/arch_exceptions.c
@@ -8,7 +8,7 @@
  *      ARMv7-M exception handlers.
  */
 
-#include "exceptions.h"
+#include <arch_exceptions.h>
 
 #include <fmw_cmsis.h>
