+++ arch/none/host/src/arch_main.c
