+++ doc/config.dxy
@@ -122,6 +122,7 @@ INPUT                  = $(TOP_DIR)/readme.md \
                          $(DOC_DIR)/cmsis.md \
                          $(DOC_DIR)/glossary.md \
                          $(DOC_DIR)/deferred_response_architecture.md \
+                         $(DOC_DIR)/architecture_support.md \
                          $(FWK_DIR)/include \
                          $(FWK_DIR)/test/fwk_test.h \
                          $(MODULE_INCLUDES) \
